+++ sktime/distances/__init__.py
@@ -1,0 +1,2 @@
+
+from .elastic_cython import *
